Moved integration tests out of code files + semantic improvments
