+++ comms/tests/connection/connection.rs
@@ -32,7 +32,7 @@ use tari_comms::connection::{
     NetAddress,
 };
 
-use tari_comms::test_support::factories::{self, Factory};
+use crate::support::factories::{self, Factory};
 
 #[test]
 fn inbound_receive_timeout() {
