+++ comms/tests/connection_manager/mod.rs
@@ -19,45 +19,5 @@
 //  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
 //  WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 //  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-use super::{peer::PeersFactory, Factory, FactoryError};
-
-use crate::{
-    peer_manager::{Peer, PeerManager},
-    types::{CommsDataStore, CommsPublicKey},
-};
-
-pub fn create() -> PeerManagerFactory {
-    PeerManagerFactory::default()
-}
-
-#[derive(Default)]
-pub struct PeerManagerFactory {
-    peers_factory: PeersFactory,
-    peers: Option<Vec<Peer<CommsPublicKey>>>,
-}
-
-impl PeerManagerFactory {
-    factory_setter!(with_peers_factory, peers_factory, PeersFactory);
-
-    factory_setter!(with_peers, peers, Option<Vec<Peer<CommsPublicKey>>>);
-}
-
-impl Factory for PeerManagerFactory {
-    type Object = PeerManager<CommsPublicKey, CommsDataStore>;
-
-    fn build(self) -> Result<Self::Object, FactoryError> {
-        let pm = PeerManager::<CommsPublicKey, CommsDataStore>::new(None)
-            .map_err(|err| FactoryError::BuildFailed(format!("Failed to build peer manager: {:?}", err)))?;
-
-        let peers = self
-            .peers
-            .or(self.peers_factory.build().ok())
-            .ok_or(FactoryError::BuildFailed("Failed to build peers".into()))?;
-        for peer in peers {
-            pm.add_peer(peer)
-                .map_err(|err| FactoryError::BuildFailed(format!("Failed to build peer manager: {:?}", err)))?;
-        }
-        Ok(pm)
-    }
-}
+// mod establisher;
+mod manager;
