+++ comms/tests/mod.rs
@@ -24,5 +24,6 @@
 extern crate lazy_static;
 
 mod connection;
+mod connection_manager;
 mod control_service;
 mod support;
