+++ comms/tests/support/mod.rs
@@ -20,5 +20,12 @@
 //  WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 //  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+#[macro_use]
+mod macros;
+
 pub mod comms_patterns;
 pub mod node_identity;
+
+pub mod factories;
+pub mod helpers;
+pub mod makers;
